squash: small refactor
