+++ workspaces/orchestrator/plugins/orchestrator-backend/src/helpers/filterBuilder.ts
@@ -425,3 +425,36 @@ function getGraphQLOperator(operator: FieldFilterOperatorEnum): string {
 function nonSecureRandomAlphaNumeric() {
   return randomBytes(8).toString('hex');
 }
+
+// For nested filters, there might be more than one filter for the same field
+// so we need to group them by the field and then combine the nested filters into an array
+export function processFilters(filter: Filter[]): Filter[] {
+  if (filter && 'filters' in filter) {
+    const grouped = (filter.filters as Filter[]).reduce<
+      Record<string, Filter[]>
+    >((acc: Record<string, Filter[]>, item: Filter) => {
+      if ('field' in item) {
+        acc[item.field] ??= [];
+        acc[item.field].push(item);
+      }
+      return acc;
+    }, {});
+
+    const newFilters: Filter[] = [];
+
+    for (const [field, filtersForField] of Object.entries(grouped)) {
+      if (filtersForField.length > 1) {
+        const nested = filtersForField
+          .filter((f): f is NestedFilter => 'nested' in f)
+          .map(f => f.nested);
+        newFilters.push({ field, nested } as unknown as NestedFilter);
+      } else {
+        newFilters.push(filtersForField[0]);
+      }
+    }
+
+    filter.filters = newFilters;
+  }
+
+  return filter;
+}
